feat(pipeline): Merge results node
Merge the results from the pipeline state into a single dictionary.

This node is used to merge the results of the extract_* nodes into a single dictionary,
since they are executed in parallel and the results are stored in different keys in the state.

By merging the results into a single dictionary, the next nodes in the pipeline can access
all the results in a easy way.